integerfield for sap id

+++ schedule/models.py
@@ -1,7 +1,7 @@
 from django.db import models
 
 class Student(models.Model):
-    sap_id = models.BigAutoField(primary_key=True)  # Long integer, Primary Key
+    sap_id = models.IntegerField(primary_key=True)  # Long integer, Primary Key
     roll_no = models.CharField(max_length=4)
     name = models.CharField(max_length=255)  # Varchar
     branch_id = models.IntegerField()  # Foreign Key placeholder (can be linked to another model later)
